requirements.txt
opencv-python
numpy
matplotlib
Pillow
pillow-heif